Adding pagination to genericAdminRoutes, refactoring search option functions into middleware
